separated out DataClasses from Shape2SAS main script; should be the end of refactoring for now
