+++ src/sas/sascalc/shape2sas/Typing.py
@@ -1,6 +1,7 @@
 import numpy as np
-from typing import Optional, Tuple, List, Any
+from typing import Tuple, List
 
+Vectors = List[List[float]]
 Vector2D = Tuple[np.ndarray, np.ndarray]
 Vector3D = Tuple[np.ndarray, np.ndarray, np.ndarray]
 Vector4D = Tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray]
